litte more styling on menu handlbars page

+++ public/js/app.js
@@ -1,5 +1,5 @@
 $(document).ready(function() {
-  $("#menu-form").hide();
+  //deletes a reservation//
   $("#deleteRes").on("click", function() {
     var id = $(this).data("id");
     $.ajax({
@@ -9,19 +9,22 @@ $(document).ready(function() {
       location.reload();
     });
   });
-  $("#deleteMenu").on("click", function() {
+  //deletes a menu item//
+  $(document).on("click", "#deleteMenu", function() {
     var id = $(this).data("id");
-    $.ajax({
-      method: "DELETE",
-      url: "/api/menu/" + id
-    }).then(function() {
-      location.reload();
-    });
+    var r = confirm("Are you sure you want to delete this item?");
+    if (r === true) {
+      $.ajax({
+        method: "DELETE",
+        url: "/api/menu/" + id
+      }).then(function() {
+        location.reload();
+      });
+    } else {
+      return false;
+    }
   });
-  $("#add").on("click", function() {
-    $("#menu-form").show();
-    $("#add").hide();
-  });
+  //adds a menu item//
   $("#submit").on("click", function() {
     event.preventDefault();
     var newItem = {
@@ -31,9 +34,7 @@ $(document).ready(function() {
       ingredients: $("#ingredients")
         .val()
         .trim(),
-      category: $("#category")
-        .val()
-        .trim(),
+      category: $("#category").val(),
       image: $("#image")
         .val()
         .trim(),
@@ -41,8 +42,19 @@ $(document).ready(function() {
         .val()
         .trim()
     };
-    $.post("/api/menu", newItem, function() {
-      location.reload();
-    });
+    if (
+      newItem.name === "" ||
+      newItem.ingredients === "" ||
+      newItem.ingredients === "" ||
+      newItem.image === "" ||
+      newItem.image === "" ||
+      newItem.price === ""
+    ) {
+      alert("Oops! Make sure everything is filled out.");
+    } else {
+      $.post("/api/menu", newItem, function() {
+        location.reload();
+      });
+    }
   });
 });
